Implement batching rules for some unary ops
This PR implements batching rules for some unary ops. In particular, it
implements the batching rules for the unary ops that take a single
tensor as input (and nothing else).

The batching rule for a unary op is:
(1) grab the physical tensor straight out of the BatchedTensor
(2) call the unary op
(3) rewrap the physical tensor in a BatchedTensor

Test Plan:
- new tests `pytest test/test_vmap.py -v -k "Operators"`

ghstack-source-id: c405ffd324870490e0348a950ee00783f2474473
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43059